v1/requirements.txt
Pillow==8.0.1
pytesseract==0.3.6
selenium==3.141.0
urllib3==1.26.0
